Fix unicode handling in get_random_string
get_random_string would often fail in decoding the result of
os.urandom. It would fail because the result of urandom is a bunch
of random bytes and not valid text (i.e., not a valid utf-8
representation).

Turns out that it's unnecessary to decode the result of os.urandom.
It returns six.binary_type whether on py2 or py3, and
binascii.hexlify() takes a six.binary_type as the argument.

binascii.hexlify() returns six.binary_type and we want
get_random_string() to return six.text_type, so convert.

Also, fixed the docstrings to use """.

Closes-Bug: 1514981
Change-Id: I60c31f2c47614beb3ed3c143d7815fdc15f5a0e0