fix notes appearing in releases older than they should
Use topological sort instead of date sort so that release notes appear
in the release where they are merged, rather than the one closest to the
same commit time, which might be much older.

In order to support listing tags on versions of git with different
output formats, we also limit the tags we care about to things that look
like version numbers (some sequence of digits, dots, and maybe "a" or
"b" for alpha & beta).

Change-Id: I95278eab25f331262a4c9eeac2260ca7290b5421
Closes-Bug: #1522153