add earliest_version option to scanner
Add an option to the scanner to tell it not to include all of the
history, but to stop at a specific version (inclusive). This will allow
us to configure release series history pages to only show versions that
are part of that series (otherwise, scanning stable/newton will include
mitaka releases from the master branch, for example).

Change-Id: I53b4b95e13c99d0a19f53e2f3e836ffe67428211
Signed-off-by: Doug Hellmann <doug@doughellmann.com>