add missing reporoot arg to 'new' command
In some of the earlier reorganization, the reporoot argument became a
required parameter. Add it as an argument to the new command with the
same default as elsewhere. It is the last argument, so it can be left
off and existing scripts using the new command will continue to work.

Change-Id: I8033f5d896201dba394346fd19e77f7c11408d6c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>