add configuration option to not stop at branch base
The previous commit changes the default behavior to always stop scanning
at the base of a branch. This change adds a configuration option to
allow that behavior to be disabled, so that revisions along the history
of the branch prior to the point where it diverged from master can be
included. The new default behavior established in the previous commit is
not changed.

Change-Id: I2c4968e1291c1b7d268896cfbb79e320d4085bce
Signed-off-by: Doug Hellmann <doug@doughellmann.com>