refactor existing implementation into a class
As part of the re-implementation work using dulwich I want to create a
Scanner class, to replace some of the module-level functions in the
scanner module now. This patch moves existing code around into something
like what I would expect that API to look like, without changing the
implementation details.

The class API is still subject to change in the future, but this should
make it a little easier to review those implementation changes.

Change-Id: If92f6ff5bce955bbdcb748cee52ad71210ae313b
Signed-off-by: Doug Hellmann <doug@doughellmann.com>