use dulwich to determine the tags on a branch
Replace the git commands for determine the tags on a branch with calls
to dulwich.

Because the Scanner class scans for the tags when it is created and
caches the data, we need to update the tests to instantiate the Scanner
after the git repo is created and the test commits are applied. The main
API entry point into reno is the Loader, and that is not changed because
the Loader does not instantiate a Scanner until it knows it needs it.

Change-Id: Icf7c7a47a768175a7cb31ab374aca4fa4b44a5cb
Signed-off-by: Doug Hellmann <doug@doughellmann.com>