use dulwich to implement get_notes_by_version
Replace the use of git-log with a traversal of the dulwich history. The
topo ordering in dulwich does not match the git command line output, so
we have our own in _topo_traversal(). I'll try to upstream that
separately.

Use the _aggregate_changes() method from the previous commit to turn a
given commit into the set of changes to make to the note history.

The topo sort in dulwich relied too much on the date ordering, which
means the entries were sometimes produced in the right order and
sometimes the wrong order.  The issue is better exposed when the commits
have more significant time differences, so add a sleep to the tests
after committing and after merging to ensure that repo changes have
distinct timestamps.

Remove the skip from the test that exposed the issue with the git
porcelain output format change to show that the new implementation
passes the test.

Since we're no longer scanning the git log history, we no longer need
the TAG_RE regex.

Change-Id: Ibfa1249f9d56a9e27b5a9f435ac8033f4bae24b6
Signed-off-by: Doug Hellmann <doug@doughellmann.com>