use dulwich's tree traversal to look up repo contents
dulwich.objects.Tree has a method lookup_path() that returns the
permissions and sha associated with a name of a thing under the
Tree (either another Tree instance or a Blob containing a file). This
change replaces the logic we have in our RenoRepo subclass for
traversing Trees with calls to lookup_path(), while still retaining
separate methods for handling the return value appropriately based on
context.

Change-Id: I29a91fa31313918e3dee9e41340b21a55cb8fa90
Signed-off-by: Doug Hellmann <doug@doughellmann.com>