fix logic for deciding when to stop scanning a branch
When scanning a given branch, we don't want to stop at the base of the
current branch, we want to stop at the base of the *previous* branch on
master. That pulls in the full history of the current branch from where
the older branch diverged.

Given this history:

  (master)
     |
   1.0.0
     | \  (stable/a)
     |  \
     |   1.0.1
   2.0.0
     |
   2.0.1
     | \ (stable/b)
     |  \
     |   2.0.2
   2.1.0

The notes for stable/b branch should include versions 2.0.2, 2.0.1, and
2.0.0 (which is on master) but not 1.0.0 which is the previous release.

Change-Id: If1feddadc1a8e24b163667cd84f5b9e098951c69
Signed-off-by: Doug Hellmann <doug@doughellmann.com>