allow tracking branch names when the branches only exist on origin
Fix a problem with branch references so that it is now possible to use a
local tracking branch name when the branch only exists on the 'origin'
remote. For example, this allows references to 'stable/ocata' when there
is no local branch with that name but there is an 'origin/stable/ocata'
branch.

Change-Id: If11a0ff10ad3b1dfc82f99e6f68684ec9268af26
Signed-off-by: Doug Hellmann <doug@doughellmann.com>