Support repodir config files
Not everyone wants to use build release notes separately from their main
documentation. For these users, having a 'notes' directory inside the
'releasenotes' directory is unnecessary. However, this also means we
must be able to move the config file out of the 'releasesnotes'
directory to avoid it being picked up as a release note.

Make this possible by adding support for a 'reno.yaml' file in the root
directory of the project.

Sadly it is not possible to apply this change to reno itself - doing so
would cause the files to be picked up as belonging to the current
release - but other projects can benefit from this.

Change-Id: Ie96103b85d70592dd766e5174784b992fe7782c5