deal with non-unique UIDs
When we find multiple files added with the same UID in one patch, ignore
them and emit a warning. That's bad input data, and we don't want to use
them.

Allow multiple files with the same UID to be deleted in a patch to
support cleaning up existing situations where the add case was not
caught properly.

Related-Bug: #1688042
Change-Id: I37fee0660ff541677d26770818764f7de2a2d863
Signed-off-by: Doug Hellmann <doug@doughellmann.com>