ignore null-merges
OpenStack used to use null-merges to bring final release tags from
stable branches back into the master branch. This confuses the regular
traversal because it makes that stable branch appear to be part of
master and/or the later stable branch. Update the scanner so that when
it hit one of those merge commits, it skips it and take the first parent
so it continues to traverse the branch being scanned.

Change-Id: I90722a3946f691e8f58a52e68ee455d6530f047a
Closes-Bug: #1695057
Signed-off-by: Doug Hellmann <doug@doughellmann.com>