loader: Extract cache filename from config object
The only places that this is called, we already have a
reno.config.Config instance available. Simply use that instead.

We also include a note of the 'conf.notespath' property, because the odd
behavior of this caught me out for a bit, and call 'os.path.normpath',
because the default path included redundant up-level references.

Change-Id: I58948cd8fad55d29bd30c65653630bd466259cdc