update ref search logic for zuulv3 CI layout
If the reference points explicitly to the origin remote, but that
remote isn't present (as it won't be when zuul configures the repo in
CI), then we want the shortened form of the reference. We put this
option last in the list because we want the more explicit name to be
used when someone is running reno locally with a more standard git
configuration.

Change-Id: I1fef83e95b0a0605db7e59676561bf1396e799e6
Signed-off-by: Doug Hellmann <doug@doughellmann.com>