ignore changes until the file is added within the scanned range
Ignore changes to a note file until the scanner encounters the git
operation that adds the file. This ensures that if a file is modified
on master when it should be modified on another branch the note is not
incorporated into the notes for the next release from master.

Change-Id: I4e41c482695e93ebb5a73866c432636201a7534f
Fixes-Bug: #1682796
Signed-off-by: Doug Hellmann <doug@doughellmann.com>