add an internal sphinx extension to show the configuration defaults
Rather than relying on contributors to add configuration data to the
usage page by hand, extract the information from the help text for the
configuration options.

Merge some of the existing documentation text with the old comments
that were turned into help text (and remove the comment markers left
behind in the last patch).

Change-Id: Ic89e6f0a083648d332ae81d20b2adbf59e4207ce
Signed-off-by: Doug Hellmann <doug@doughellmann.com>