improve output messages
Change the order of initialization so the logging module is configured
before the configuration file is read. This lets us log debug messages
from the config module.

Add some information from the config module about which configuration
file is loaded, and in verbose mode which files we're looking for.

Add some information to the scanner output to show whether
collapse_pre_releases is set to true.

Change-Id: Ia78bc63e2c29de2ac8771b46397a63639012bb90
Addresses-Bug: #1743202
Signed-off-by: Doug Hellmann <doug@doughellmann.com>