support scanning closed stable branches
The scanner was stopping too soon when reviewing the history of a
branch for which the previous branch had been "closed" by deleting the
branch and tagging it with an -eol tag. This fix treats closed -eol
branches the same way as open stable branches, using new configuration
options to allow projects that use different naming conventions to
have the same benefits.

Change-Id: I8024929a2a95e00df48ce56939d54c1569fe18c5
Fixes-Bug: #1746076
Signed-off-by: Doug Hellmann <doug@doughellmann.com>