Make section titles have stable anchor links
Currently some sections produce anchors like "#id1", which will change
over time as new releases are made. Instead, put in explicit references
so that each version gets an anchor based on the version number, and so
that each heading inside of the version has one based on the section
title and the release version.

Each anchor needs to be prefixed either with the title, if given, or
with another string as some people include reno inside of other
sphinx documentation and these will produce global refererences for the
sphinx build. Also, since it's one global set of anchors, each version
needs to be prefixed by title (or 'relnotes') and the section title
because otherwise in cases like reno's own docs where some versions are
included in the output twice, sphinx will produce conflicts.

Change-Id: Ia6bdaffa6d0ae286542fbb7ae12613be56bdb326