Collapse Unreleased and Mainline sections
The current release notes produced have an 'Unreleased' section that
includes the releases back through 2.0.0, then mainline that includes
those release notes as well as 1.5.0 through 0.1.0.

Squish the two so that there is just one copy of each of them. Remove
the explicit mention of origin/master and the earliest-version and just
let the normal reno logic apply.

Change-Id: I02bb2c69b8dd9e866e4efa3437e9c09c33488914