include the branch name in anchors to make them more unique
If the same version number appears in multiple output pages, sphinx
reports a warning because the anchors generated by the formatter are
the same. This patch adds the branch name to the anchor to make it
more unique, since we should only be scanning a given branch one time
in each release notes build.

See
http://logs.openstack.org/05/564405/1/check/build-openstack-releasenotes/b5fad95/job-output.txt.gz#_2018-04-27_06_02_12_077967
for an example of a failed build.

Change-Id: I0b02e7eb319c95f885fc494977b356af71370970
Signed-off-by: Doug Hellmann <doug@doughellmann.com>