Switch to use stestr for unit test
According to Openstack summit session [1], stestr is maintained project
to which all Openstack projects should migrate.  So we should switch to
stestr.

This has the side effect of no longer reporting coverage by default for
unit tests. This isn't a big deal given that we have a 'cover' target
for just this purpose.

[1] https://etherpad.openstack.org/p/YVR-python-pti

Change-Id: I85545c7b4c410c310026911448a7c3847c665afb