Use unicode for debug string
On Python2, the "line" variable here is a unicode string, but the
LOG.debug statement is trying to format it with a bytes string; this
leads to a conversion error when the line has unicode characters in
it.  Ensure the logging format string is a unicode object to avoid
this.

Change-Id: I9948ffcea3fd10c4b6695593fa2a42731b44c02f