only override config values from command line if they are actually set
If we set defaults for the query configuration options and then pass
the namespace created by parsing the command line options to the
config object it cannot tell the difference between values set on the
command line and defaults. That means we may override settings in the
configuration file with defaults from the command line parser.

This change sets all of the defaults for command line options to None
as a sentinel value and then updates the Config class to ignore
command line option values of None. It also adds tests to verify that
we get the expected override behavior with boolean and string options.

Change-Id: I1c9ce668b5e5c372d1c861bcae6e6de05a8ebc0c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>