Handle Windows compatibility
Some of the library functions we call for file path handling require the
calling code to handle the differences in path separators for Windows.
This adds special handling where this was found, though it is possible
there are some other code paths that have not been identified yet.

Story: 2006549
Task: 36632

Change-Id: I0d74222f711724f3aaccc226a9ec4b5f31cf283e
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>