Add extension metadata to sphinxext
This commit adds returned extension metadata to the reno sphinx
extension. Sphinx enables extensions to return metadata by returning a
dictionary from the setup function [1]. This is mostly used for setting
version information and specifying how the extension behaves when
running in a sphinx parallel build. The metadata returned by the reno
sphinx extension here is used to tell sphinx implicitly that it is not
parallel read safe which avoid a warning by sphinx build saying the
extension doesn't specify. But, it also adds a version returned so that
other sphinx extensions can depend on the reno extension.

[1] https://www.sphinx-doc.org/en/master/extdev/index.html#extension-metadata

Change-Id: I80a671a98901eb45ecb732d699400a76326b4cf2
Task: 37082
Story: 2006712