Keep scanning master branch after first shared commit
If no earliest version is provided for a branch, reno will scan commits
on the branch in reverse order, attempting to find a common ancestor
with the master branch. Currently, the first patch this finds must be
the tagged one or it will quit early. There's no particular reason to do
this. Instead, we can keep going for a little longer. This is helpful
for situations where we haven't branched off straight after the tag.

Change-Id: Ic0ed1305cf21deb6abea0991ec604d7f9b821622
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>