Handle empty config files
Current code was assuming if a reno.yaml file was found that it could
load that file to read config options. While probably a safe assumption
in the majority of cases, someone that isn't aware of this could modify
their config file to remove settings, leaving an empty or comment-only
file that when read does not actually load any yaml that can be used.

This adds protection for that to make reno does not try to use an empty
file.

Change-Id: I0ae0bcc93a79bbf66735833b654964b494485ee1
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>