Set parallel_read_safe to True
This commit sets the parallel_read_safe extension metadata attribute to
True. This was previously set to False as a precaution and to avoid the
warning about the implicit default in change
I80a671a98901eb45ecb732d699400a76326b4cf2. However, in subsequent
testing running a sphinx build in parallel with the reno extension does
not cause any errors or issues in the built documentation and trying to
run in parallel with this value set to false raises a warning (which is
fatal if -W is set). This commit makes this change so that people can
use sphinx-build's parallel build option (-j) and -W at the same time.

Change-Id: I5a3b5f11d94c2a29c04582b4dde813d67a0efbc1