Handle repositories with a default branch other than master
This commit adds a new configuration option `default_branch`, which is
used to specify the default/base branch a repository is using.
Previously reno assumed the default branch was always named 'master'
which causes commands such as 'reno lint' to fail when it is run on a
repository that uses a different default branch name such as 'main'.
This config option gives users an option to specify which branch should
be used for cases like that.

Change-Id: I08f258206290fdc1dcb9fcfc078e30384667c925