Close files opened by Scanner, Loader
We're seeing noise like the following in doc builds:

  /foo/.tox/docs/lib/python3.9/site-packages/docutils/statemachine.py:707:
  ResourceWarning: unclosed file <_io.BufferedReader name='/foo/.git/objects/pack/pack-dd69481843ca1b7377f2f109b0022221437aca20.pack'>
    if not hasattr(pattern, 'match'):
  ResourceWarning: Enable tracemalloc to get the object allocation traceback

Enable tracemalloc for Sphinx gives us the 'Loader' as the root cause:

  $ python -W all::ResourceWarning -X tracemalloc=100 -m sphinx.cmd.build ...
  ...
  File "/foo/.tox/docs/lib/python3.9/site-packages/reno/sphinxext.py", lineno 114
    ldr = loader.Loader(conf)
  File "/foo/.tox/docs/lib/python3.9/site-packages/reno/loader.py", lineno 63
    self._load_data()
  ...

(you could also use PYTHONTRACEMALLOC and PYTHONWARNINGS env vars)

Following this thread, it appears 'reno.scanner.Scanner' creates an
instance of 'dulwich.repo.Repo', however, it fails to close it [1]. The
'reno.loader.Loader' uses 'Scanner', meaning this also leaves around
open files.

The solution is simple: add a 'close()' method to both the 'Scanner' and
'Loader', and provide the necessary '__enter__' and '__exit__' magic
methods to use it as a context manager, like the 'Repo' object itself
supports.

[1] https://www.dulwich.io/docs/api/dulwich.repo.html#dulwich.repo.Repo.close

Change-Id: I0b9776f431cf902a9ace5d52961eb77caaae8eaa
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>