Respect EOM tag for branches in unmaintained status
With the new unmaintained process [1], branches are being tagged with
EOM before being removed.

This behaviour breaks reno, making it struggle to find respective origin
and leads to releasenotes job failure.

In order to overcome the issue we add eom tag to the list of expected ones
when we can not find the refferenced branch.

With that, EOL tag will have prescedence over EOM one.

[1] https://governance.openstack.org/tc/resolutions/20230724-unmaintained-branches.html

Change-Id: Ic641af5aa65add9751b49fb47691a98f80a7c5a8